* Compiler warnings in XspfLoader.

+++ src/libtomahawk/utils/XspfLoader.cpp
@@ -241,7 +241,7 @@ XSPFLoader::gotBody()
         m_title = m_overrideTitle;
 
     bool shownError = false;
-    for ( int i = 0; i < tracklist.length(); i++ )
+    for ( int i = 0; i < (int)tracklist.length(); i++ )
     {
         QDomNode e = tracklist.at( i );
